Moved out the custom hash map class to its own file
